replace box with arc mutex

+++ src/constants.rs
@@ -1,2 +1,2 @@
-pub const INTEGER_BYTES: usize = 4;
+pub const INTEGER_BYTES: i32 = 4;
 pub const LOG_FILE: &str = "simpledb.log";
