+++ src/tx/recovery/set_int_record.rs
@@ -47,20 +47,20 @@ impl std::fmt::Display for SetIntRecord {
 
 impl SetIntRecord {
     pub fn new_from_page(p: Page) -> Result<Self, String> {
-        let tpos = INTEGER_BYTES;
+        let tpos = INTEGER_BYTES as usize;
         let txnum = p.get_int(tpos)?;
 
-        let fpos = tpos + INTEGER_BYTES;
+        let fpos = tpos + INTEGER_BYTES as usize;
         let filename = p.get_string(fpos)?;
 
         let bpos = fpos + Page::max_length(filename.len());
         let blknum = p.get_int(bpos)?;
         let blk = BlockId::new(filename, blknum);
 
-        let opos = bpos + INTEGER_BYTES;
+        let opos = bpos + INTEGER_BYTES as usize;
         let offset = p.get_int(opos)? as usize;
 
-        let vpos = opos + INTEGER_BYTES;
+        let vpos = opos + INTEGER_BYTES as usize;
         let val = p.get_int(vpos)?;
 
         Ok(Self {
@@ -78,13 +78,13 @@ impl SetIntRecord {
         offset: i32,
         val: i32,
     ) -> Result<i32, String> {
-        let tpos = INTEGER_BYTES;
-        let fpos = tpos + INTEGER_BYTES;
+        let tpos = INTEGER_BYTES as usize;
+        let fpos = tpos + INTEGER_BYTES as usize;
         let bpos = fpos + Page::max_length(blk.file_name().len());
-        let opos = bpos + INTEGER_BYTES;
-        let vpos = opos + INTEGER_BYTES;
+        let opos = bpos + INTEGER_BYTES as usize;
+        let vpos = opos + INTEGER_BYTES as usize;
 
-        let mut p = Page::new_from_blocksize(vpos + INTEGER_BYTES);
+        let mut p = Page::new_from_blocksize(vpos + INTEGER_BYTES as usize);
         p.set_int(0, log_record::SETINT)?;
         p.set_int(tpos, txnum)?;
         p.set_string(fpos, &blk.file_name())?;
